Add componenr Register and test, fixes in test register page
