+++ package.json
@@ -19,7 +19,7 @@
     "react-scripts": "3.4.4",
     "react-spinkit": "^3.0.0",
     "seamless-immutable": "^7.1.4",
-    "typescript": "^3.9.7",
+    "typescript": "^4.0",
     "wolox-equalizer": "0.0.3"
   },
   "devDependencies": {
