+++ src/app/components/RegisterForm/i18n.ts
